+++ Monster.h
@@ -18,7 +18,7 @@ class Monster
 public:
 
 
-    Monster(std::string name, int health, int damage, float attackCooldown);
+    Monster(std::string name, int health, int damage, double attackCooldown);
     Monster() {};
 	///This function parse a Monster from a file.
 	///\param s The file name we parse.
@@ -58,8 +58,10 @@ public:
     ///Attacks the enemy.
     ///\param enemy The monster we attack
     void getAttacked(const Monster& enemy);
+	///Operator overwrite.
     friend bool operator==(const Monster character1, const Monster character2);
-    friend bool operator!=(const Monster character1, const Monster character2);
+    ///Operator overwrite.
+	friend bool operator!=(const Monster character1, const Monster character2);
 private:
     virtual void Attack(Monster& enemy);
